Refactor: Improve component organization in Checkout page
This commit refactors the Checkout page to improve
component organization.  The `PaymentTypeCard`
component is moved to a dedicated subdirectory
within the `src/pages/landing` folder. This
change enhances maintainability and readability
of the codebase.  Additionally, unnecessary
imports in `PricingCard.tsx` have been removed to
reduce bundle size and improve performance.